+++ src/aws_central_infrastructure/identity_center/lib/__init__.py
@@ -1,10 +1,11 @@
-from .lib import LOW_RISK_ADMIN_PERM_SET_CONTAINER
-from .lib import VIEW_ONLY_PERM_SET_CONTAINER
-from .lib import AwsAccountInfo
-from .lib import AwsLogicalWorkload
-from .lib import AwsSsoPermissionSet
-from .lib import AwsSsoPermissionSetAccountAssignments
-from .lib import AwsSsoPermissionSetContainer
-from .lib import DefaultWorkloadPermissionAssignments
 from .lib import User
+from .lib import UserInfo
 from .lib import create_read_state_inline_policy
+from .permissions import LOW_RISK_ADMIN_PERM_SET_CONTAINER
+from .permissions import VIEW_ONLY_PERM_SET_CONTAINER
+from .permissions import AwsAccountInfo
+from .permissions import AwsLogicalWorkload
+from .permissions import AwsSsoPermissionSet
+from .permissions import AwsSsoPermissionSetAccountAssignments
+from .permissions import AwsSsoPermissionSetContainer
+from .permissions import DefaultWorkloadPermissionAssignments
